64 bit cleanup
There are a lot of changes in this patch so it should be tested well...
In general I changed the following types of things:

  If the internals used a specific type eg. uint32_t then I exposed that
  thru the interface.

  If the internals used a generic type (int, long, etc) then I changed
  them to ssize_t UNLESS all useage of the variable where unsigned in
  which case I used size_t and fixed the new collisions.

  I used VERY few casts and only in cases where I could follow the logic
  thru to the cast being safe and used code guards otherwise.

+++ libzt/adt/zt_bstream.h
@@ -17,10 +17,10 @@ struct zt_bstream {
 
 /* predicates */
 extern int zt_bstream_is_empty(zt_bstream_t bs);
-extern int zt_bstream_truncate(zt_bstream_t bs);
+extern size_t zt_bstream_truncate(zt_bstream_t bs);
 extern void zt_bstream_set_data(zt_bstream_t bs, char *data, int len, char copy);
 extern void zt_bstream_set_array(zt_bstream_t bs, zt_array_t array, size_t offt);
-extern int zt_bstream_rewind(zt_bstream_t bs);
+extern size_t zt_bstream_rewind(zt_bstream_t bs);
 
 extern zt_bstream_t  zt_bstream_new(void);
 extern zt_bstream_t  zt_bstream_clone(zt_bstream_t bs);
